[core, Lua] Add additional Lua bindings for locking/reserving items

+++ src/map/packet_system.cpp
@@ -2904,7 +2904,7 @@ void SmallPacket0x04D(map_session_data_t* const PSession, CCharEntity* const PCh
                     if (!PChar->UContainer->IsSlotEmpty(deliverySlotID))
                     {
                         CItem* PItem = PChar->UContainer->GetItem(deliverySlotID);
-                        if (PItem->isSent())
+                        if (PItem && PItem->isSent())
                         {
                             ret = sql->Query("DELETE FROM delivery_box WHERE charid = %u AND box = 2 AND slot = %u LIMIT 1;", PChar->id, deliverySlotID);
                             if (ret != SQL_ERROR && sql->AffectedRows() == 1)
@@ -6937,7 +6937,7 @@ void SmallPacket0x0FA(map_session_data_t* const PSession, CCharEntity* const PCh
 
         // Update installed furniture placement orders
         // First we place the furniture into placed items using the order number as the index
-        std::array<CItemFurnishing*, MAX_CONTAINER_SIZE* 2> placedItems = { nullptr };
+        std::array<CItemFurnishing*, MAX_CONTAINER_SIZE * 2> placedItems = { nullptr };
         for (auto safeContainerId : { LOC_MOGSAFE, LOC_MOGSAFE2 })
         {
             CItemContainer* PContainer = PChar->getStorage(safeContainerId);
@@ -6961,7 +6961,7 @@ void SmallPacket0x0FA(map_session_data_t* const PSession, CCharEntity* const PCh
         }
 
         // Update the item's order number
-        for (int32 i = 0; i < MAX_CONTAINER_SIZE * 2; ++i)
+        for (int32 i = 0; i < MAX_CONTAINER_SIZE* 2; ++i)
         {
             // We can stop updating the order numbers once we hit an empty order number
             if (placedItems[i] == nullptr)
@@ -7763,6 +7763,10 @@ void SmallPacket0x106(map_session_data_t* const PSession, CCharEntity* const PCh
 
     CItemContainer* PBazaar         = PTarget->getStorage(LOC_INVENTORY);
     CItemContainer* PBuyerInventory = PChar->getStorage(LOC_INVENTORY);
+    if (PBazaar == nullptr || PBuyerInventory == nullptr)
+    {
+        return;
+    }
 
     if (PChar->id == PTarget->id || PBuyerInventory->GetFreeSlotsCount() == 0)
     {
@@ -7872,6 +7876,10 @@ void SmallPacket0x109(map_session_data_t* const PSession, CCharEntity* const PCh
 {
     TracyZoneScoped;
     CItemContainer* PStorage = PChar->getStorage(LOC_INVENTORY);
+    if (PStorage == nullptr)
+    {
+        return;
+    }
 
     for (uint8 slotID = 1; slotID <= PStorage->GetSize(); ++slotID)
     {
@@ -7898,7 +7906,17 @@ void SmallPacket0x10A(map_session_data_t* const PSession, CCharEntity* const PCh
     uint8  slotID = data.ref<uint8>(0x04);
     uint32 price  = data.ref<uint32>(0x08);
 
-    CItem* PItem = PChar->getStorage(LOC_INVENTORY)->GetItem(slotID);
+    auto* PStorage = PChar->getStorage(LOC_INVENTORY);
+    if (PStorage == nullptr)
+    {
+        return;
+    }
+
+    CItem* PItem = PStorage->GetItem(slotID);
+    if (PItem == nullptr)
+    {
+        return;
+    }
 
     if (PItem->getReserve() > 0)
     {
